Moved cylindrical_isolation into isolation_functions sub-package
