+++ halotools/mock_observables/isolation_functions/__init__.py
@@ -3,4 +3,5 @@ isolation criteria functionality.
 """
 
 from .spherical_isolation import spherical_isolation
+from .cylindrical_isolation import cylindrical_isolation
